[STRM-1456] [LYFT] Add kinesis and kafka wrappers to python sdk
