+++ sdks/python/apache_beam/io/lyft/__init__.py
@@ -1,0 +1,1 @@
+"""A package defining Lyft-internal sources and sinks"""
