+++ sdks/python/custom-source-example.py
@@ -1,120 +1,11 @@
 import logging
-import json
 import sys
 import apache_beam as beam
 from apache_beam.io import ReadFromText
+from apache_beam.io.lyft.kafka import FlinkKafkaInput
+from apache_beam.io.lyft.kinesis import FlinkKinesisInput
 from apache_beam.options.pipeline_options import PipelineOptions
 from apache_beam.runners.portability import portable_runner
-
-from apache_beam.transforms.ptransform import PTransform
-from apache_beam.transforms.window import GlobalWindows
-from apache_beam.transforms.core import Windowing
-from apache_beam import pvalue
-
-from google.protobuf import wrappers_pb2
-
-
-class FlinkKafkaInput(PTransform):
-    """Custom transform that wraps a Flink Kafka consumer - only works with the portable Flink runner."""
-    consumer_properties = {'bootstrap.servers' : 'localhost:9092'}
-    topic = None
-
-    def expand(self, pbegin):
-        assert isinstance(pbegin, pvalue.PBegin), (
-                'Input to transform must be a PBegin but found %s' % pbegin)
-        return pvalue.PCollection(pbegin.pipeline)
-
-    def get_windowing(self, inputs):
-        return Windowing(GlobalWindows())
-
-    def infer_output_type(self, unused_input_type):
-        return bytes
-
-    def to_runner_api_parameter(self, context):
-        assert isinstance(self, FlinkKafkaInput), \
-            "expected instance of CustomKafkaInput, but got %s" % self.__class__
-        assert self.topic is not None, "topic not set"
-        assert len(self.consumer_properties) > 0, "consumer properties not set"
-        return ("lyft:flinkKafkaInput",
-            json.dumps({ 'topic' : self.topic, 'properties' : self.consumer_properties})
-                )
-
-    @staticmethod
-    @PTransform.register_urn("lyft:flinkKafkaInput", None)
-    #@PTransform.register_urn("custom:kafkaInput", wrappers_pb2.BytesValue)
-    def from_runner_api_parameter(spec_parameter, unused_context):
-        print "spec: " + spec_parameter
-        instance = FlinkKafkaInput()
-        payload = json.loads(spec_parameter)
-        instance.topic = payload['topic']
-        instance.consumer_properties = payload['properties']
-        return instance
-
-    def with_topic(self, topic):
-        self.topic = topic
-        return self
-
-    def set_kafka_consumer_property(self, key, value):
-        self.consumer_properties[key] = value;
-        return self
-
-    def with_bootstrap_servers(self, bootstrap_servers):
-        return self.set_kafka_consumer_property('bootstrap.servers', bootstrap_servers)
-
-    def with_group_id(self, group_id):
-        return self.set_kafka_consumer_property('group.id', group_id)
-
-
-class FlinkKinesisInput(PTransform):
-    """Custom transform that wraps a Flink Kinesis consumer - only works with the portable Flink runner."""
-    consumer_properties = {'aws.region' : 'us-east-1', 'flink.stream.initpos' : 'TRIM_HORIZON'}
-    stream = None
-
-    def expand(self, pbegin):
-        assert isinstance(pbegin, pvalue.PBegin), (
-                'Input to transform must be a PBegin but found %s' % pbegin)
-        return pvalue.PCollection(pbegin.pipeline)
-
-    def get_windowing(self, inputs):
-        return Windowing(GlobalWindows())
-
-    def infer_output_type(self, unused_input_type):
-        return bytes
-
-    def to_runner_api_parameter(self, context):
-        assert isinstance(self, FlinkKinesisInput), \
-            "expected instance of FlinkKinesisInput, but got %s" % self.__class__
-        assert self.stream is not None, "topic not set"
-        assert len(self.consumer_properties) > 0, "consumer properties not set"
-        return ("lyft:flinkKinesisInput",
-                json.dumps({ 'stream' : self.stream, 'properties' : self.consumer_properties})
-                )
-
-    @staticmethod
-    @PTransform.register_urn("lyft:flinkKinesisInput", None)
-    def from_runner_api_parameter(spec_parameter, unused_context):
-        print "spec: " + spec_parameter
-        instance = FlinkKinesisInput()
-        payload = json.loads(spec_parameter)
-        instance.stream = payload['stream']
-        instance.consumer_properties = payload['properties']
-        return instance
-
-    def with_stream(self, stream):
-        self.stream = stream
-        return self
-
-    def set_consumer_property(self, key, value):
-        self.consumer_properties[key] = value;
-        return self
-
-    # to use Kinesalite
-    def with_endpoint(self, endpoint, accessKey, secretKey):
-        self.consumer_properties.pop('aws.region', None) # cannot have both region and endpoint
-        self.set_consumer_property('aws.endpoint', endpoint)
-        self.set_consumer_property('aws.credentials.provider.basic.accesskeyid', accessKey)
-        self.set_consumer_property('aws.credentials.provider.basic.secretkey', secretKey)
-        return self
 
 
 if __name__ == "__main__":
@@ -136,10 +27,10 @@ if __name__ == "__main__":
 
   with beam.Pipeline(runner=runner, options=pipeline_options) as p:
     (p
-        #| 'Create' >> beam.Create(['hello', 'world', 'world'])
-        #| 'Read' >> ReadFromText("gs://dataflow-samples/shakespeare/kinglear.txt")
-        #| 'Kafka' >> FlinkKafkaInput().with_topic('beam-example').with_bootstrap_servers('localhost:9092').with_group_id('beam-example-group')
-        | 'Kinesis' >> FlinkKinesisInput().with_stream('beam-example').with_endpoint('http://localhost:4567', 'fakekey', 'fakesecret')
+        # | 'Create' >> beam.Create(['hello', 'world', 'world'])
+        # | 'Read' >> ReadFromText("gs://dataflow-samples/shakespeare/kinglear.txt")
+        | 'Kafka' >> FlinkKafkaInput().with_topic('beam-example').with_bootstrap_servers('localhost:9092').with_group_id('beam-example-group')
+        # | 'Kinesis' >> FlinkKinesisInput().with_stream('beam-example').with_endpoint('http://localhost:4567', 'fakekey', 'fakesecret')
         #| 'Split' >> (beam.FlatMap(lambda x: re.findall(r'[A-Za-z\']+', x))
         #                .with_output_types(unicode))
         #| 'PairWithOne' >> beam.Map(lambda x: (x, 1))
